Refactor (io): Unified the return type of read interfaces to int
- Changed the return value of AbstractFilterReader's read method from size_t to int
- Modified AbstractInputStream's read method to return std::byte instead of size_t
- Updated the read method signatures of AbstractReader and its derived classes to uniformly return int
- Adjusted the read method implementation of BufferedInputStream to adapt to the new return value type
- Optimized ByteArrayInputStream and CharArrayReader to add closed status checks
- Improved boundary condition handling and exception safety of various read methods
- Added the isClosed method to all input stream classes to support closed status queries
- Enhanced the IAppendable interface to support more types of character append operations
- Refactored the IBoostSerializable interface to simplify serialization access control

+++ common/filesystem/io/interface/IReadable.hpp
@@ -15,7 +15,7 @@ namespace fox
 
         /// @brief Read one byte from the input source.
         /// This method reads and returns the next byte of data from the input source.
-        /// @return The byte value read, or -1 if the end of the stream has been reached
-        virtual auto read() -> size_t = 0;
+        /// @return The byte value read as an unsigned integer, or -1 if the end of the stream has been reached
+        virtual auto read() -> int = 0;
     };
 }
